Add gamma index to RTTB
Summary: This is the addition of the gamma index as an index to RTTB. It has been lying around for a while and I have actually done very little of the coding work on it. I would still like to merge it (with minimal further effort), so as to not lose the invested work.

Test Plan: Build and use as custom RTTB for a MITK build based on D845 and play around with the Gamma Index plugin

Reviewers: O2 RTTB, floca

Reviewed By: O2 RTTB, floca

Maniphest Tasks: T27843

Differential Revision: https://phabricator.mitk.org/D844